+++ tests/testthat/test_icap.R
@@ -1,5 +1,5 @@
 # context("ICAP")
 # test_that('ICAP files available', {
-#   expect_equal(download(url = url_icap_zip(TRUE), destfile = tempfile(), quiet = TRUE), 0L)
-#   expect_equal(download(url = url_icap_zip(FALSE), destfile = tempfile(), quiet = TRUE), 0L)
+#   expect_equal(downloader::download(url = url_icap_zip(TRUE), destfile = tempfile(), quiet = TRUE), 0L)
+#   expect_equal(downloader::download(url = url_icap_zip(FALSE), destfile = tempfile(), quiet = TRUE), 0L)
 # })