+++ NEWS.md
@@ -3,6 +3,7 @@
 - Added ability to source trade data from SDRs (DTCC, Bloomberg and CME) (#7) 
 - Can now specify `asset_class` in `get_bsef_data()`. Defaults to downloading data for all asset classes
 - Deprecated ICAP SEF capabilities (unstable)
+- Removed packrat-sourced download() method in favour of downloader package
 
 # Version 0.1